Analyst can Save new resource URLs [#160921702]

+++ src/main/java/mil/af/dgs1sdt/blackpearl/resource/ResourceRepository.java
@@ -2,6 +2,6 @@ package mil.af.dgs1sdt.blackpearl.resource;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResourceRepository extends JpaRepository<Resource, Integer> {
+public interface ResourceRepository extends JpaRepository<Resource, Long> {
 
 }
